+++ app/v2/users/views.py
@@ -1,9 +1,30 @@
 """app/v1/users/views.py"""
 from functools import wraps
 import jwt
+import re
 from flask import request, jsonify, session
 from . import users_api
 from app.models import User
+from app.jwt import Auth
+from app.database.conn import dbcon
+
+jwt_auth = Auth()
+userObject = User()
+
+
+def is_admin_loggedin():
+    """ check if a user is an admin logged in"""
+    header = request.headers.get('authorization')
+    token = header.split(" ")[1]
+    token = jwt.decode(token, 'SECRET_KEY', algorithms=['HS256'])
+    user_id = token['userid']
+    conn = dbcon()
+    cur = conn.cursor()
+    cur.execute("SELECT * FROM tbl_users WHERE userid=%(userid)s AND userRole=%(role)s",\
+    {'userid': user_id, 'role': 'admin'})
+    if cur.rowcount > 0:
+        return True
+    return False
 
 
 def token_required(f):
@@ -11,19 +32,33 @@ def token_required(f):
     @wraps(f)
     def decorated(**kwargs):
         """decorator"""
-        try:
-            session['token']
-        except BaseException:
-            session['token'] = None
+        header = request.headers.get('Authorization')
 
-        if session['token'] is None:
+        if header is None:
+            return jsonify({"message": "Authorization header missing"}), 403
+        elif header.split(" ")[1] is None:
+            return jsonify({"message": "Token missing"}), 403
+        token = header.split(" ")[1]
+
+        try:
+            if jwt_auth.in_blacklist(token) is False:
+                try:
+                    data = jwt.decode(token, 'SECRET_KEY', algorithms=['HS256'])
+                except jwt.ExpiredSignatureError:
+                    # token expired blacklist token
+                    jwt_auth.blacklist(token)
+                    return jsonify({'message': 'Logged out. Please login and update token'}), 401
+                except jwt.InvalidTokenError:
+                    # token invalid blacklist token
+                    jwt_auth.blacklist(token)
+                    return jsonify({'message': 'Invalid Token. Please login'}), 403
+            else:
+                return jsonify({'message': 'Token blacklisted. Please login'}), 401
+        except BaseException:
             return jsonify({'message': 'Please login'}), 401
 
         return f(**kwargs)
     return decorated
-
-
-userObject = User()
 
 
 def validate_data_signup(data):
@@ -50,12 +85,6 @@ def validate_data_signup(data):
         # check if userphone empty
         elif data["userphone"] == "":
             return "userphone required"
-        # check if userRole has spaces
-        elif " " in data["userRole"]:
-            return "userRole should be one word, no spaces"
-        # check if userRole is empty
-        elif data["userRole"] == "":
-            return "userRole required"
         elif len(data['password'].strip()) < 5:
             return "Password should have atleast 5 characters"
         # check if the passwords match
@@ -101,12 +130,10 @@ def reg():
         username = data['username']
         userphone = data['userphone']
         password = data['password']
-        userRole = data['userRole']
         response = userObject.create_user(
             username,
             userphone,
-            password,
-            userRole)
+            password)
         return response
     return jsonify({"message": res}), 400
 
@@ -121,51 +148,59 @@ def login():
         password = data['password']
         response = userObject.login(username, password)
         return response
-    return jsonify({"message": res}), 401
+    return jsonify({"message": res}), 403
 
 
 @users_api.route('/users', methods=["GET"])
 @token_required
 def users():
     """get all users"""
-    data = userObject.get_users()
-    return data
+    if is_admin_loggedin() is True:
+        data = userObject.get_users()
+        return data
+    return jsonify({
+        "message": "You dont have admin priviledges."}), 401
 
 
 @users_api.route('/users/<int:id>', methods=["GET", "DELETE", "PUT"])
 @token_required
 def users_verbs(id):
     """run http verbs"""
-    if request.method == "GET":
-        data = userObject.get_specific_user(id)
-        return data
-    elif request.method == "PUT":
-        data = request.get_json()
-        res = validate_data_signup(data)
+    if is_admin_loggedin() is True:
+        if request.method == "GET":
+            data = userObject.get_specific_user(id)
+            return data
+        elif request.method == "PUT":
+            data = request.get_json()
+            res = validate_data_signup(data)
 
-        if res == "valid":
-            username = data['username']
-            userphone = data['userphone']
-            password = data['password']
-            userRole = data['userRole']
-            response = userObject.update_user(
-                id,
-                username,
-                userphone,
-                password,
-                userRole)
+            if res == "valid":
+                username = data['username']
+                userphone = data['userphone']
+                password = data['password']
+                response = userObject.update_user(
+                    id,
+                    username,
+                    userphone,
+                    password)
+                return response
+            return jsonify({"message": res}), 400
+        elif request.method == "DELETE":
+            response = userObject.delete_user(id)
             return response
-        return jsonify({"message": res}), 400
-    elif request.method == "DELETE":
-        response = userObject.delete_user(id)
-        return response
+    return jsonify({
+        "message": "You dont have admin priviledges."}), 401
 
 
 @users_api.route('/auth/logout')
+@token_required
 def logout():
     """ Method to logout user."""
-    if 'token' in session:
-        session.clear()
-        return jsonify({"message": "Succeffully logout."})
-    return jsonify({
-        "message": "Not logged in."}), 200
+    header = request.headers.get('authorization')
+    if header is None:
+        return jsonify({"message": "No token provided"}), 400
+    token = header.split(" ")[1]
+    res = jwt_auth.blacklist(token)
+    if res is True:
+        return jsonify({"message": "Succefuly logout."}), 200
+    print(res)
